Move toolbar product links to eth-common plugin
- Allows linking to Alethio products in full explorer and lite explorer
