+++ packages/eth-extended/config.sample.json
@@ -33,9 +33,5 @@
 
     "UNCLE_API_URL_MASK": "https://blk-api.ethstats.io/api/v3/uncle/%s",
 
-    "alethioApiUrl": "https://api.aleth.io/docs",
-    "alethioMonitoringUrl": "https://monitoring.aleth.io",
-    "alethioReportsUrl": "https://reports.aleth.io",
-    "ethstatsUrl": "https://ethstats.io",
-    "companyUrl": "https://company.aleth.io"
+    "ethstatsUrl": "https://ethstats.io"
 }
