+++ src/app/eth-extended/ethExtendedPlugin.ts
@@ -28,7 +28,6 @@ import { accountContractModule } from "./module/account/contract/accountContract
 import { accountDetailsModule } from "./module/account/accountDetails/accountDetailsModule";
 import { avgTimeInPoolChartModule } from "./module/dashboard/charts/avgTimeInPoolChartModule";
 import { propagationChartModule } from "./module/dashboard/charts/propagationChartModule";
-import { ethstatsModule } from "./module/toolbar/ethstatsModule";
 import { BlockAvgTxTimeInPoolAdapter } from "./adapter/block/BlockAvgTxTimeInPoolAdapter";
 import { blockDetailsModule } from "./module/block/blockDetails/blockDetailsModule";
 import { unclePage } from "./page/uncle/unclePage";
@@ -40,23 +39,7 @@ import { txDetailsModule } from "app/eth-extended/module/tx/txDetails/txDetailsM
 import { BlockBasicInfoAdapter } from "app/shared/adapter/block/BlockBasicInfoAdapter";
 import { AlethioAdapterType } from "app/shared/adapter/AlethioAdapterType";
 import { dashboardPage } from "app/eth-extended/page/dashboard/dashboardPage";
-import { alethioApiModule as toolbarApiModule } from "app/eth-extended/module/toolbar/alethioApiModule";
-import {
-    alethioMonitoringModule as toolbarMonitoringModule
-} from "app/eth-extended/module/toolbar/alethioMonitoringModule";
-import { alethioReportsModule as toolbarReportsModule } from "app/eth-extended/module/toolbar/alethioReportsModule";
-import { alethioApiModule as topbarApiModule } from "app/eth-extended/module/topbar/alethioApiModule";
-import {
-    alethioMonitoringModule as topbarMonitoringModule
-} from "app/eth-extended/module/topbar/alethioMonitoringModule";
-import { alethioReportsModule as topbarReportsModule } from "app/eth-extended/module/topbar/alethioReportsModule";
 import { uncleByHashContextType } from "app/eth-extended/context/uncleByHashContextType";
-import {
-    alethioCompanyModule as toolbarAlethioCompanyModule
-} from "app/eth-extended/module/toolbar/alethioCompanyModule";
-import {
-    alethioCompanyModule as topbarAlethioCompanyModule
-} from "app/eth-extended/module/topbar/alethioCompanyModule";
 
 const ethExtendedPlugin: IPlugin = {
     init(config, api, logger, publicPath) {
@@ -113,22 +96,6 @@ const ethExtendedPlugin: IPlugin = {
         api.addDataAdapter("adapter://aleth.io/pendingPoolInfo", new PendingPoolInfoAdapter(dataSource));
         api.addDataAdapter("adapter://aleth.io/propagationInfo", new PropagationInfoAdapter(dataSource));
         api.addDataAdapter("adapter://aleth.io/ethNodesInfo", new EthNodesInfoAdapter(dataSource));
-
-        api.addModuleDef("module://aleth.io/toolbar/ethstats", ethstatsModule(config.ethstatsUrl));
-
-        api.addModuleDef("module://aleth.io/toolbar/alethioApi", toolbarApiModule(config.alethioApiUrl));
-        api.addModuleDef("module://aleth.io/topbar/alethioApi", topbarApiModule(config.alethioApiUrl));
-
-        api.addModuleDef("module://aleth.io/toolbar/alethioMonitoring",
-            toolbarMonitoringModule(config.alethioMonitoringUrl));
-        api.addModuleDef("module://aleth.io/topbar/alethioMonitoring",
-            topbarMonitoringModule(config.alethioMonitoringUrl));
-
-        api.addModuleDef("module://aleth.io/toolbar/alethioReports", toolbarReportsModule(config.alethioReportsUrl));
-        api.addModuleDef("module://aleth.io/topbar/alethioReports", topbarReportsModule(config.alethioReportsUrl));
-
-        api.addModuleDef("module://aleth.io/toolbar/company", toolbarAlethioCompanyModule(config.companyUrl));
-        api.addModuleDef("module://aleth.io/topbar/company", topbarAlethioCompanyModule(config.companyUrl));
 
         api.addPageDef("page://aleth.io/dashboard", dashboardPage);
         api.addModuleDef("module://aleth.io/dashboard/avgTimeInPoolChart", avgTimeInPoolChartModule);
